backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
python-multipart
alembic
rapidfuzz
celery[redis]
redis
slowapi
pytest
